requirements.txt
pandas==2.2.2
minio==7.2.8
pyarrow==17.0.0
apache-airflow==2.10.1
pytest==8.3.3  # For unit testing
plotly==5.24.1
dbt-core==1.7.2
dbt-postgres==1.7.2
apache-airflow-providers-docker==3.14.0